# Conflicts:
#   client/src/MobicoopBundle/Community/Controller/CommunityController.php

+++ client/src/MobicoopBundle/User/Controller/UserController.php
@@ -113,6 +113,8 @@ class UserController extends AbstractController
      */
     public function userSignUp(UserManager $userManager, Request $request, TranslatorInterface $translator)
     {
+        $this->denyAccessUnlessGranted('register');
+
         $user = new User();
         $address = new Address();
         $error = false;
@@ -236,6 +238,9 @@ class UserController extends AbstractController
             if (!$homeAddress) {
                 $homeAddress = new Address();
             }
+
+            $this->denyAccessUnlessGranted('address_update_self', $user);
+            
             
             $address=json_decode($data->get('homeAddress'), true);
             $homeAddress->setAddressCountry($address['addressCountry']);
@@ -307,6 +312,7 @@ class UserController extends AbstractController
     public function userProfileAvatarDelete(ImageManager $imageManager, UserManager $userManager)
     {
         $user = clone $userManager->getLoggedUser();
+        $this->denyAccessUnlessGranted('update', $user);
         $imageId = $user->getImages()[0]->getId();
         $imageManager->deleteImage($imageId);
 
@@ -370,6 +376,7 @@ class UserController extends AbstractController
      */
     public function userPasswordRecovery()
     {
+        $this->denyAccessUnlessGranted('login');
         return $this->render('@Mobicoop/user/passwordRecovery.html.twig', []);
     }
 
@@ -380,6 +387,7 @@ class UserController extends AbstractController
      */
     public function userPasswordForRecovery(UserManager $userManager, Request $request)
     {
+        $this->denyAccessUnlessGranted('login');
         if ($request->isMethod('POST')) {
             $data = json_decode($request->getContent(), true);
 
@@ -421,6 +429,8 @@ class UserController extends AbstractController
             $data = json_decode($request->getContent(), true);
             
             $user = $userManager->findByPwdToken($token);
+
+            $this->denyAccessUnlessGranted('password', $user);
 
             if (!empty($user)) {
                 $user->setPassword($data["password"]);
@@ -510,6 +520,7 @@ class UserController extends AbstractController
     public function userMessageDirectThreadsList(UserManager $userManager, InternalMessageManager $internalMessageManager)
     {
         $user = $userManager->getLoggedUser();
+        $this->denyAccessUnlessGranted('messages', $user);
         $threads = $userManager->getThreadsDirectMessages($user);
         return new Response(json_encode($threads));
     }
@@ -520,6 +531,7 @@ class UserController extends AbstractController
     public function userMessageCarpoolThreadsList(UserManager $userManager, InternalMessageManager $internalMessageManager)
     {
         $user = $userManager->getLoggedUser();
+        $this->denyAccessUnlessGranted('messages', $user);
         $threads = $userManager->getThreadsCarpoolMessages($user);
         return new Response(json_encode($threads));
     }
@@ -527,8 +539,10 @@ class UserController extends AbstractController
     /**
      * Get direct messages threads
      */
-    public function userMessageThread($idMessage, InternalMessageManager $internalMessageManager)
+    public function userMessageThread($idMessage, InternalMessageManager $internalMessageManager, UserManager $userManager)
     {
+        $user = $userManager->getLoggedUser();
+        $this->denyAccessUnlessGranted('messages', $user);
         $completeThread = $internalMessageManager->getThread($idMessage, DataProvider::RETURN_JSON);
         return new Response(json_encode($completeThread));
     }
